moved all files and set up useReducer, deleted data.js and UserProvider/userContext
