--- umts/src/components/providers/UserProvider.js
@@ -1,15 +1,0 @@
-import React, { useState } from 'react';
-
-import { userContext } from '../../contexts/userContext';
-
-const UserProvider = ({ children }) => {
-  const [user, setUser] = useState({});
-
-  return (
-    <userContext.Provider value={{ user, setUser }}>
-      {children}
-    </userContext.Provider>
-  );
-};
-
-export default UserProvider;
