+++ umts/src/contexts/tech/techContext.js
@@ -1,0 +1,5 @@
+import { createContext } from 'react';
+
+const TechContext = createContext();
+
+export default TechContext;
