+++ umts/src/contexts/tech/techReducer.js
@@ -1,0 +1,6 @@
+export default (state, action) => {
+  switch (action.type) {
+    default:
+      return state;
+  }
+};
