--- umts/src/contexts/techContext.js
@@ -1,3 +1,0 @@
-import { createContext } from 'react';
-
-export const techContext = createContext();
