--- umts/src/data.js
@@ -1,10 +1,0 @@
-export const itemData = [
-  {
-    title: 'Canon EOS M5 Mirrorless Digital Camera',
-    description:
-      '24.2MP APS-C CMOS Sensor, DIGIC Image Processor, 2.36m-Dot EVF, Touch and Drag AF Control, 3.2" 1.62m-Dot Tilting Touchscreen LCD',
-    category: 'Camera',
-    price_per_hr: '$30',
-    price_per_day: '$150'
-  }
-];
